add line-item class to accounting namespace; deprecate old classes
