+++ src/XeroPHP/Models/Accounting/Prepayment/LineItem.php
@@ -2,206 +2,10 @@
 
 namespace XeroPHP\Models\Accounting\Prepayment;
 
-use XeroPHP\Remote;
-use XeroPHP\Models\Accounting\TrackingCategory;
+use XeroPHP\Models\Accounting;
 
-class LineItem extends Remote\Model
-{
-    /**
-     * Description needs to be at least 1 char long. A line item with just a description (i.e no unit
-     * amount or quantity) can be created by specifying just a <Description> element that contains at least
-     * 1 character.
-     *
-     * @property string Description
-     */
-
-    /**
-     * LineItem Quantity.
-     *
-     * @property string Quantity
-     */
-
-    /**
-     * Lineitem unit amount. By default, unit amount will be rounded to two decimal places. You can opt in
-     * to use four decimal places by adding the querystring parameter unitdp=4 to your query. See the
-     * Rounding in Xero guide for more information.
-     *
-     * @property float UnitAmount
-     */
-
-    /**
-     * See Accounts.
-     *
-     * @property string AccountCode
-     */
-
-    /**
-     * Used as an override if the default Tax Code for the selected <AccountCode> is not correct – see
-     * TaxTypes.
-     *
-     * @property string TaxType
-     */
-
-    /**
-     * The tax amount is auto calculated as a percentage of the line amount (see below) based on the tax
-     * rate. This value can be overriden if the calculated <TaxAmount> is not correct.
-     *
-     * @property float TaxAmount
-     */
-
-    /**
-     * If you wish to omit either of the <Quantity> or <UnitAmount> you can provide a LineAmount and Xero
-     * will calculate the missing amount for you.
-     *
-     * @property float LineAmount
-     */
-
-    /**
-     * Optional Tracking Category – see Tracking. Any LineItem can have a maximum of 2 <TrackingCategory>
-     * elements.
-     *
-     * @property TrackingCategory[] Tracking
-     */
-
-    /**
-     * Get the resource uri of the class (Contacts) etc.
-     *
-     * @return string
-     */
-    public static function getResourceURI()
-    {
-        return 'LineItems';
-    }
-
-    /**
-     * Get the root node name.  Just the unqualified classname.
-     *
-     * @return string
-     */
-    public static function getRootNodeName()
-    {
-        return 'LineItem';
-    }
-
-    /**
-     * Get the guid property.
-     *
-     * @return string
-     */
-    public static function getGUIDProperty()
-    {
-        return '';
-    }
-
-    /**
-     * Get the stem of the API (core.xro) etc.
-     *
-     * @return string|null
-     */
-    public static function getAPIStem()
-    {
-        return Remote\URL::API_CORE;
-    }
-
-    /**
-     * Get the supported methods.
-     */
-    public static function getSupportedMethods()
-    {
-        return [
-        ];
-    }
-
-    /**
-     * Get the properties of the object.  Indexed by constants
-     *  [0] - Mandatory
-     *  [1] - Type
-     *  [2] - PHP type
-     *  [3] - Is an Array
-     *  [4] - Saves directly.
-     *
-     * @return array
-     */
-    public static function getProperties()
-    {
-        return [
-            'Description' => [false, self::PROPERTY_TYPE_STRING, null, false, false],
-            'Quantity' => [false, self::PROPERTY_TYPE_STRING, null, false, false],
-            'UnitAmount' => [false, self::PROPERTY_TYPE_FLOAT, null, false, false],
-            'AccountCode' => [false, self::PROPERTY_TYPE_STRING, null, false, false],
-            'TaxType' => [false, self::PROPERTY_TYPE_ENUM, null, false, false],
-            'TaxAmount' => [false, self::PROPERTY_TYPE_FLOAT, null, false, false],
-            'LineAmount' => [false, self::PROPERTY_TYPE_FLOAT, null, false, false],
-            'Tracking' => [false, self::PROPERTY_TYPE_OBJECT, 'Accounting\\TrackingCategory', true, false],
-        ];
-    }
-
-    public static function isPageable()
-    {
-        return false;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDescription()
-    {
-        return $this->_data['Description'];
-    }
-
-    /**
-     * @return string
-     */
-    public function getQuantity()
-    {
-        return $this->_data['Quantity'];
-    }
-
-    /**
-     * @return float
-     */
-    public function getUnitAmount()
-    {
-        return $this->_data['UnitAmount'];
-    }
-
-    /**
-     * @return string
-     */
-    public function getAccountCode()
-    {
-        return $this->_data['AccountCode'];
-    }
-
-    /**
-     * @return string
-     */
-    public function getTaxType()
-    {
-        return $this->_data['TaxType'];
-    }
-
-    /**
-     * @return float
-     */
-    public function getTaxAmount()
-    {
-        return $this->_data['TaxAmount'];
-    }
-
-    /**
-     * @return float
-     */
-    public function getLineAmount()
-    {
-        return $this->_data['LineAmount'];
-    }
-
-    /**
-     * @return Remote\Collection|TrackingCategory[]
-     */
-    public function getTracking()
-    {
-        return $this->_data['Tracking'];
-    }
-}
+/**
+ * @deprecated 2.0.5 use \XeroPHP\Models\Accounting\LineItem instead
+ */
+class LineItem extends Accounting\LineItem
+{}
