+++ src/XeroPHP/Remote/Model.php
@@ -492,8 +492,6 @@ abstract class Model implements ObjectInterface, \JsonSerializable, \ArrayAccess
         }
 
         trigger_error(sprintf("Undefined property %s::$%s.\n", __CLASS__, $property));
-
-        
     }
 
     /**
@@ -513,8 +511,6 @@ abstract class Model implements ObjectInterface, \JsonSerializable, \ArrayAccess
         }
 
         trigger_error(sprintf("Undefined property %s::$%s.\n", __CLASS__, $property));
-
-        
     }
 
     protected function propertyUpdated($property, $value)
